refactor: MatchResult 패키지 변경
