+++ src/test/java/blackjack/model/referee/MatchResultTest.java
@@ -1,6 +1,7 @@
 package blackjack.model.referee;
 
 import blackjack.model.card.Score;
+import blackjack.model.result.MatchResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
